Configure path to MBL includes used during compile
With this change, we expect to see the test named "Spawn is able to
compile a Modelica model that uses external functions" pass. This will
be notable because it compiles
"Buildings.ThermalZones.EnergyPlus.Validation.ThermalZone.OneZoneOneYear"
which is a Modelica model that incorportes Spawn components. The
remaining key ingredient was to enable the hooks that allow the Modelica compiler
to call back to the (embedded) C compiler when encountering external objects (which
spawn models use).

Simply put, compiling a Modelica model to FMU is as simple as:

`spawn --compile
"Buildings.ThermalZones.EnergyPlus.Validation.ThermalZone.OneZoneOneYear"`

The necessary dependencies including Modelica compiler and Clang are bottled up in spawn(exe).

There are caveats:

1. This is all only wired up for Linux.
2. The configuration management works in the developer context, but in
an installed production context there is still install and path configuration
work to be done.
