+++ test/paths.in.cpp
@@ -20,11 +20,3 @@ fs::path example_fmu_path() {
   return "$<TARGET_FILE:example_fmu_lib>";
 }
 
-fs::path project_source_dir() {
-  return "${PROJECT_SOURCE_DIR}";
-}
-
-fs::path project_binary_dir() {
-  return "${PROJECT_BINARY_DIR}";
-}
-
