+++ util/paths.hpp
@@ -6,8 +6,14 @@
 namespace spawn {
   // todo Move idd paths, etc into here
   fs::path exe();
+
   fs::path exedir();
   
+  fs::path mbl_home_dir();
+
+  fs::path project_source_dir();
+  
+  fs::path project_binary_dir();
 }
